Fix usenet_groups query
The `$backfilldays` variable may either be an int or string depending
on the value of the selected option is in the tmux settings -> backfill
(third dropdown). If the selected option is "Safe Backfill day",
`$backfilldays` will be an int. If it's "Days per Group",
`$backfilldays` will be a string. With `%d`, the string value is
replaced with a 0, causing the result to always be an empty result set.