Escape Message-ID during update binaries
I ran misc/update/update_binaries.php manually because I noticed a lot
of missing parts that weren't going away and noticed sql errors on the
INSERT IGNORE. It turns out that some of the message IDs had single
quotes in them that inhibited the query from executing properly.

An example I found is: 200420092300004385%But-i-really@dont'care.com
(oddly enough, the single quote isn't in the word don't, but after...)

semi-sanitizing the data through the `addslashes` function takes care of
the problem -- I'm not aware if there's a more elegant method of
sanitizing that is available. Please let me know if there is and I'll
update.